coffull-docker/api/requirements.txt
Django>=2.2.0, <2.3.0
psycopg2
djangorestframework
django-cors-headers